// src/app/shared/types/api-output-types/task-management/bire-task-measurement-point-table-dto.interface.ts
export interface BireTaskMeasurementPointTableDTO {
  counterCode?: string;
  unitCode?: string;
  counterRemarks?: string;
  taskCode?: string;
  taskVersion?: string;
  sequenceNumber?: number;
  groupCode?: string;
  formula?: string;
  toolPartNumber?: string;
  minValue?: number;
  maxValue?: number;
  description?: string;
  additionalDescription?: string;
}